fix: rename backup components
